requirements.txt
Flask==0.12
requests==2.18.4
nltk==3.2.4
indicoio==1.1.1
numpy==1.13.1
urllib3==1.22
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
Werkzeug==0.12.1
itsdangerous==0.24
Jinja2==2.8
click==6.7
six==1.10.0
mock==1.3.0
pbr==3.1.1
pillow==4.2.1
olefile==0.44
futures==3.1.1
gunicorn==19.7.1
vaderSentiment==2.5
newspaper3k==0.2.2
Image==1.5.15
